requirements.txt
mysql-connector-python>=8.0.0
python-json-logger>=2.0.0
jsonschema>=4.0.0 